telegram_bot_no_miniapp/requirements.txt
python-telegram-bot
httpx
flask
requests
